pyproject.toml
[project]
name = "v1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.5",
    "langchain>=0.3.27",
    "ollama>=0.5.1",
    "openai>=1.98.0",
    "pillow>=11.3.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.54.1",
]
